pyproject.toml
[project]
name = "aa-sdr-generator"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "authlib>=1.7.1",
    "openpyxl>=3.1.5",
    "pydantic>=2.13.3",
    "pydantic-settings>=2.14.0",
    "pyyaml>=6.0.3",
    "requests>=2.33.1",
]

[dependency-groups]
dev = [
    "pytest>=9.0.3",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
